REFACTOR :: 인풋 컴포넌트 분리
